Update README w/ PoC-4 and Testnet faucet
The README was a bit out-of-date.  PoC-4 is out (and PoC-3 no longer works with the Alexander Testnet, at least for me).  I worked with Logan to come up with the new instructions for building and tested them on OS X.

Additionally, I added a link to the new Testnet Faucet as an alternative way to get Testnet DOTs.